All shape file data that is read in and represented as agents is done so using type specific contexts and geographies.
This means that the readShapefileWithType function is not needed and I can revert to using the original readShapefile function (which does the same thing as readShapefile with type).

Could do with some more though on how to organise and manage different contexts and geographies but for now keeping them separate is best.